+++ source/Relativity.Testing.Framework.Api/Services/BatchService.cs
@@ -11,17 +11,23 @@ namespace Relativity.Testing.Framework.Api.Services
 		private readonly IGetAllWorkspaceEntitiesStrategy<Batch> _getAllWorkspaceEntitiesStrategy;
 		private readonly IBatchQueryStrategy _batchQueryStrategy;
 		private readonly IBatchAssignToUserStrategy _batchAssignToUserStrategy;
+		private readonly IBatchCheckinStrategy _batchCheckinStrategy;
+		private readonly IBatchCheckoutStrategy _batchCheckoutStrategy;
 
 		public BatchService(
 			IGetWorkspaceEntityByIdStrategy<Batch> getWorkspaceEntityByIdStrategy,
 			IGetAllWorkspaceEntitiesStrategy<Batch> getAllWorkspaceEntitiesStrategy,
 			IBatchQueryStrategy batchQueryStrategy,
-			IBatchAssignToUserStrategy batchAssignToUserStrategy)
+			IBatchAssignToUserStrategy batchAssignToUserStrategy,
+			IBatchCheckinStrategy batchCheckinStrategy,
+			IBatchCheckoutStrategy batchCheckoutStrategy)
 		{
 			_getWorkspaceEntityByIdStrategy = getWorkspaceEntityByIdStrategy;
 			_getAllWorkspaceEntitiesStrategy = getAllWorkspaceEntitiesStrategy;
 			_batchQueryStrategy = batchQueryStrategy;
 			_batchAssignToUserStrategy = batchAssignToUserStrategy;
+			_batchCheckinStrategy = batchCheckinStrategy;
+			_batchCheckoutStrategy = batchCheckoutStrategy;
 		}
 
 		public Batch Get(int workspaceId, int entityId)
@@ -35,5 +41,11 @@ namespace Relativity.Testing.Framework.Api.Services
 
 		public void AssignToUser(int workspaceId, int batchId, int userId)
 			=> _batchAssignToUserStrategy.AssignToUser(workspaceId, batchId, userId);
+
+		public void Checkin(int workspaceId, int batchId, bool isCompleted)
+			=> _batchCheckinStrategy.Checkin(workspaceId, batchId, isCompleted);
+
+		public void Checkout(int workspaceId, int batchId, int userId)
+			=> _batchCheckoutStrategy.Checkout(workspaceId, batchId, userId);
 	}
 }