+++ source/Relativity.Testing.Framework.Api/Strategies/Batches/IBatchCheckinStrategy.cs
@@ -1,0 +1,7 @@
+﻿namespace Relativity.Testing.Framework.Api.Strategies
+{
+	public interface IBatchCheckinStrategy
+	{
+		void Checkin(int workspaceId, int batchId, bool isCompleted);
+	}
+}